Fix: Pre-pack F16 as F32 for serial GEMMs

+++ include/numkong/spatials/serial.h
@@ -49,11 +49,11 @@ nk_define_cross_normalized_symmetric_(euclidean, f32, serial, f32, f64, /*norm_v
                                       nk_dots_reduce_sumsq_f32_, nk_load_b256_serial_, nk_partial_load_b64x4_serial_,
                                       nk_store_b256_serial_, nk_partial_store_b64x4_serial_, 1)
 
-nk_define_cross_normalized_packed_(angular, f16, serial, f16, f16, f32, /*norm_value_type=*/f32, f32, nk_b128_vec_t,
+nk_define_cross_normalized_packed_(angular, f16, serial, f16, f32, f32, /*norm_value_type=*/f32, f32, nk_b128_vec_t,
                                    nk_dots_packed_f16_serial, nk_angular_through_f32_from_dot_serial_,
                                    nk_dots_reduce_sumsq_f16_, nk_load_b128_serial_, nk_partial_load_b32x4_serial_,
                                    nk_store_b128_serial_, nk_partial_store_b32x4_serial_, 1)
-nk_define_cross_normalized_packed_(euclidean, f16, serial, f16, f16, f32, /*norm_value_type=*/f32, f32, nk_b128_vec_t,
+nk_define_cross_normalized_packed_(euclidean, f16, serial, f16, f32, f32, /*norm_value_type=*/f32, f32, nk_b128_vec_t,
                                    nk_dots_packed_f16_serial, nk_euclidean_through_f32_from_dot_serial_,
                                    nk_dots_reduce_sumsq_f16_, nk_load_b128_serial_, nk_partial_load_b32x4_serial_,
                                    nk_store_b128_serial_, nk_partial_store_b32x4_serial_, 1)
